+++ dice-calculator/func_maximumpossibilities.h
@@ -10,7 +10,7 @@ int64_t func_maximumpossibilities(std::vector<Diceset> dicesets, int64_t size) {
 
 	// Raise all sides to the power of the amount of dice thrown and multiply them to get the maximumpossibilities
 	for (int64_t i = 0; i < size; i++) {
-		maximumpossibilities *= std::pow(dicesets[i].sides, dicesets[i].amount);
+		maximumpossibilities *= static_cast<int64_t>(std::pow(dicesets[i].sides, dicesets[i].amount));
 	}
 
 	return maximumpossibilities;
